IR: Add the VariableRegion annotation
This annotation will permit to get where a variable is produced and
where it's consumed. We only identify CPU and GPU regions for now. The
BOTH region means that the variable is used on the CPU and the GPU (for
the read region).

Also add a basic annotation placement on IR object.