[dev] Implement custom NablaIdeCrossrefProposalProvider
This provider is provided to workaround a difference of behavior between
the RCP and LSP version. In the RCP version, if a library is imported
its provided items are only displayed using their aliased version (not
the qualified on) in the proposals. In the LSP version, both the aliased
and qualified name version are displayed causing a lot of useless
entries in the content assist proposals.

Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>